fee-caching functions added, kyc functions fixed

+++ src/tradelayer/tx.h
@@ -150,7 +150,7 @@ private:
     char hash[SP_STRING_FIELD_LEN];
     int block_forexpiry;
     uint8_t tokens, ltc, natives, oracles;
-    std::vector<int> kyc_Ids; //kyc vector
+    std::vector<int64_t> kyc_Ids; //kyc vector
 
 
     // Indicates whether the transaction can be used to execute logic
